+++ asyncio_rpc/exceptions.py
@@ -1,3 +1,8 @@
+class NotReceived(Exception):
+    """
+    Message has not been recieved by anyone
+    """
+
 class WrappedException(Exception):
     """
     Exception raised when an exception raised